UTIL-3 - use long for CHS resource size
git-svn-id: https://source.sakaiproject.org/contrib/etudes/etudes-util/trunk@69179 fdecad78-55fc-0310-b1b2-d7d25cf747c9

+++ etudes-util/util/src/java/org/etudes/util/XrefHelper.java
@@ -1113,7 +1113,7 @@ public class XrefHelper
 			// get the source
 			ContentResource resource = ContentHostingService.getResource(id);
 			String type = resource.getContentType();
-			int size = resource.getContentLength();
+			long size = resource.getContentLength();
 			byte[] body = resource.getContent();
 
 			String[] parts = StringUtil.split(id, "/");
@@ -1464,7 +1464,7 @@ public class XrefHelper
 	 * @return true if there is a match, false if not.
 	 * @throws ServerOverloadException
 	 */
-	protected static boolean matches(ContentResource mr, String type, int size, byte[] body) throws ServerOverloadException
+	protected static boolean matches(ContentResource mr, String type, long size, byte[] body) throws ServerOverloadException
 	{
 		// check if the type and size match - get the cheap tests done first
 		if (!StringUtil.different(mr.getContentType(), type, true))
